requirements.txt
fastapi==0.73.0
motor==2.5.1
pydantic[email]
python-decouple==3.6
uvicorn==0.17.4
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4